moves result widgets to options.py
