--- docs/source/menpowidgets/menpofitwidgets/options/IterativeResultOptionsWidget.rst
@@ -1,13 +1,0 @@
-.. _menpowidgets-menpofitwidgets-options-IterativeResultOptionsWidget:
-
-.. currentmodule:: menpowidgets.menpofitwidgets.options
-
-IterativeResultOptionsWidget
-============================
-.. autoclass:: IterativeResultOptionsWidget
-  :members: add_render_function, remove_render_function, replace_render_function, call_render_function, set_widget_state, style, predefined_style,
-            add_displacements_function, remove_displacements_function, replace_displacements_function,
-            add_errors_function, remove_errors_function, replace_errors_function,
-            add_costs_function, remove_costs_function, replace_costs_function,
-            add_callbacks, remove_callbacks, set_visibility, add_traits, close, has_trait, observe, trait_names, traits, unobserve, unobserve_all
-  :show-inheritance:
