--- docs/source/menpowidgets/menpofitwidgets/options/ResultOptionsWidget.rst
@@ -1,10 +1,0 @@
-.. _menpowidgets-menpofitwidgets-options-ResultOptionsWidget:
-
-.. currentmodule:: menpowidgets.menpofitwidgets.options
-
-ResultOptionsWidget
-===================
-.. autoclass:: ResultOptionsWidget
-  :members: add_render_function, remove_render_function, replace_render_function, call_render_function, set_widget_state, style, predefined_style,
-            add_callbacks, remove_callbacks, set_visibility, add_traits, close, has_trait, observe, trait_names, traits, unobserve, unobserve_all
-  :show-inheritance:
