--- docs/source/menpowidgets/menpofitwidgets/options/index.rst
@@ -1,13 +1,0 @@
-.. _api-menpofitwidgets-options-index:
-
-:mod:`menpowidgets.menpofitwidgets.options`
-===========================================
-
-Independent widget classes that can be used as the main components for
-designing high-level widget functions, as the ones in :ref:`api-menpofitwidgets-base-index`.
-
-.. toctree::
-    :maxdepth: 2
-
-    ResultOptionsWidget
-    IterativeResultOptionsWidget
